// MochiR.Client/src/lib/api/types.ts
export interface components {
	schemas: {
// ...
		CreateReviewDto: {
			/** Format: int32 */
			subjectId: number;
			title: string | null;
			content: string | null;
			excerpt: string | null;
			ratings: components['schemas']['ReviewRatingDto'][] | null;
			tags: string[] | null;
		};
// ...
		LatestReviewsPageDto: {
			/** Format: int32 */
			totalCount: number;
			/** Format: int32 */
			page: number;
			/** Format: int32 */
			pageSize: number;
			items: components['schemas']['ReviewSummaryDto'][];
			nextCursor: components['schemas']['LatestReviewsCursorDto'];
			hasMore: boolean;
		} | null;
// ...
		ReviewSummaryDto: {
			/** Format: int64 */
			id: number;
			/** Format: int32 */
			subjectId: number;
			subjectName: string | null;
			userId: string;
			authorUserName: string | null;
			authorDisplayName: string | null;
			authorAvatarUrl: string | null;
			title: string | null;
			content: string | null;
			excerpt: string | null;
			excerptIsAuto: boolean;
			ratings: components['schemas']['ReviewRatingDto'][];
			status: components['schemas']['ReviewStatus'];
			/** Format: date-time */
			createdAt: string;
			tags: string[];
			/** Format: int32 */
			likeCount: number;
			isLikedByCurrentUser: boolean;
		};
// ...
		UpdateReviewDto: {
			title: string;
			content: string | null;
			excerpt: string | null;
			ratings: components['schemas']['ReviewRatingDto'][] | null;
			tags: string[] | null;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}